Rewrite of the Grails Cxf Plugin.
This is an opinionated rewrite of the Grails Cxf Plugin in preparation
for adding lots more support to it. The rewrite takes into account the
best Grails plugin practices that I know of as well as Spring-ifying
more of the plugin configuration.

Part of the rewrite was to move away from using Grails Service Artefacts
for Cxf Soap endpoints to using a specific artefact type. While the idea
of using normal services is a great one for small projects, once a
project gets larger services have a tendency to become a catch all for
code that isn't DAO and isn't presentation. Typically its better if we
create much more granular classes to do specific things (think SRP).
These new Endpoint Artefacts are autowired injectable singletons but
unlike Grails Service Artefacts are not proxies, will not create
transactional boundaries, and will not receive special treatment from
other plugins (unless supported).

Because of the use of a new Grails Artefact type the various Cxf
ServiceFactoryBeans have been simplified and the duplicate code for
creating the list of ignored methods is gone. Also spring now configures
properties of the ServiceFactoryBeans rather than doing it as part of
the constructor.

The GrailsCxfServlet has also been simplified. One of the first things I
did was to update Cxf to the most recent version, whatever issues they
had with the normal CxfServlet are gone; the class was able to migrate
back to using the CxfServlet as well as not having to lookup the spring
ApplicationContext on it's own. The GrailsCxfServlet has been boiled
down to the bean wiring that needs to be done.

New Features to Come
====================
 * More Configuration
 * WSDL First Development
 * Either integrated or plugin support for WS-Security
 * Interceptors
 * Integrated client support (WSDL First, and Code First Development)
 * A default set of Features (database logging, filesystem logging)
 * Multiplexed endpoint references

Also as always the test suite should be drastically improved and user
documentation expanded on.

I hope that the direction that this fork is taking can either be
integrated back with the main plugin or can be released as a separate
Cxf plugin.