Two New Basic Scripts
CreateEndpoint
--------------
This script can be used to generate very basic stub classes for the
Endpoint Artefacts.

WsdlToJava
----------
A script that makes use of the Cxf WSDLToJava tool to generate java code
for WSDL objects. It currently generates a Service class that can be
discarded.

Both scripts could use some TLC on argument parsing and expanding their
features. Argument parsing in Grails/Gant scripts is not what I think of
as fun.